fix(198): corrections issues identifiées en code review
- refresh_records_mv : met à jour cutoff_date après REFRESH MATERIALIZED VIEW
  pour éviter les doublons dans HybridTemperatureRecordsDataSource
- Renommer _fetch_hot_records → _fetch_records_after_cutoff (nom trompeur)
- Supprimer RecordAbsolu model (inutilisé, incohérence de type department)
- Déduplication des helpers de test (insert_station, insert_quotidienne)
  déplacés dans conftest.py
- Typo : "Parametre" → "Paramètre" dans le message d'erreur de la view
