+++ backend/weather/tests/integration/test_temperature_records_timescale_datasource.py
@@ -3,74 +3,12 @@ from __future__ import annotations
 import datetime as dt
 
 import pytest
-from django.db import connection
 
 from weather.data_sources.timescale import (
     TimescaleTemperatureRecordsDataSource,
 )
 from weather.services.temperature_records.types import TemperatureRecordsRequest
-
-# =========================
-# Helpers SQL
-# =========================
-
-
-def insert_station(code: str, name: str = "Station test", department: int = 1) -> None:
-    now = dt.datetime.now()
-
-    with connection.cursor() as cur:
-        cur.execute(
-            """
-            INSERT INTO public."Station"
-                ("createdAt", "updatedAt", "id", "nom",
-                 "departement", "frequence",
-                 "posteOuvert", "typePoste",
-                 "lon", "lat", "alt", "postePublic")
-            VALUES
-                (%(created)s, %(updated)s, %(id)s, %(name)s,
-                 %(dept)s, 'horaire',
-                 '1', 1,
-                 0.0, 0.0, 0.0, '1')
-            ON CONFLICT ("id", "frequence") DO UPDATE SET "nom" = EXCLUDED."nom"
-            """,
-            {
-                "created": now,
-                "updated": now,
-                "id": code,
-                "name": name,
-                "dept": department,
-            },
-        )
-
-
-def insert_quotidienne_full(
-    day: dt.date,
-    code: str,
-    *,
-    tx: float | None = None,
-    tn: float | None = None,
-) -> None:
-    """Insert a Quotidienne row with TX and TN columns (not just TNTXM)."""
-    with connection.cursor() as cur:
-        cur.execute(
-            """
-            INSERT INTO public."Quotidienne"
-                ("NUM_POSTE", "NOM_USUEL", "LAT", "LON", "ALTI", "AAAAMMJJ", "TX", "TN", "TNTXM")
-            VALUES
-                (%(code)s, %(name)s, 0, 0, 0, %(day)s, %(tx)s, %(tn)s, %(tntxm)s)
-            ON CONFLICT ("NUM_POSTE", "AAAAMMJJ")
-            DO UPDATE SET "TX" = EXCLUDED."TX", "TN" = EXCLUDED."TN", "TNTXM" = EXCLUDED."TNTXM"
-            """,
-            {
-                "code": code,
-                "name": f"ST {code}",
-                "day": day,
-                "tx": tx,
-                "tn": tn,
-                "tntxm": ((tx or 0) + (tn or 0)) / 2 if tx and tn else None,
-            },
-        )
-
+from weather.tests.conftest import insert_quotidienne, insert_station
 
 # =========================
 # Tests
@@ -87,9 +25,9 @@ def test_fetch_records_hot_month_happy_path():
     # 2003: 38.0 → premier record (prev_max=NULL)
     # 2019: 42.6 → 42.6 > 38.0 → record progressif
     # 2020: 35.0 → 35.0 < 42.6 → pas un record
-    insert_quotidienne_full(dt.date(2003, 7, 15), station_code, tx=38.0, tn=20.0)
-    insert_quotidienne_full(dt.date(2019, 7, 25), station_code, tx=42.6, tn=22.0)
-    insert_quotidienne_full(dt.date(2020, 7, 10), station_code, tx=35.0, tn=19.0)
+    insert_quotidienne(dt.date(2003, 7, 15), station_code, tx=38.0, tn=20.0)
+    insert_quotidienne(dt.date(2019, 7, 25), station_code, tx=42.6, tn=22.0)
+    insert_quotidienne(dt.date(2020, 7, 10), station_code, tx=35.0, tn=19.0)
 
     ds = TimescaleTemperatureRecordsDataSource()
     req = TemperatureRecordsRequest(period_type="month", type_records="hot", month=7)
@@ -115,8 +53,8 @@ def test_fetch_records_cold_month_happy_path():
 
     insert_station(station_code, "Station Cold Test", department=99)
 
-    insert_quotidienne_full(dt.date(1985, 1, 16), station_code, tx=0.0, tn=-20.5)
-    insert_quotidienne_full(dt.date(2010, 1, 7), station_code, tx=2.0, tn=-10.0)
+    insert_quotidienne(dt.date(1985, 1, 16), station_code, tx=0.0, tn=-20.5)
+    insert_quotidienne(dt.date(2010, 1, 7), station_code, tx=2.0, tn=-10.0)
 
     ds = TimescaleTemperatureRecordsDataSource()
     req = TemperatureRecordsRequest(period_type="month", type_records="cold", month=1)
@@ -141,9 +79,9 @@ def test_fetch_records_season_aggregates_across_months():
     # 2003-08-12: 40.0 → premier record estival
     # 2019-06-28: 44.0 → 44.0 > 40.0 → record progressif
     # 2019-07-25: 42.6 → 42.6 < 44.0 → pas un record
-    insert_quotidienne_full(dt.date(2003, 8, 12), station_code, tx=40.0, tn=21.0)
-    insert_quotidienne_full(dt.date(2019, 6, 28), station_code, tx=44.0, tn=25.0)
-    insert_quotidienne_full(dt.date(2019, 7, 25), station_code, tx=42.6, tn=22.0)
+    insert_quotidienne(dt.date(2003, 8, 12), station_code, tx=40.0, tn=21.0)
+    insert_quotidienne(dt.date(2019, 6, 28), station_code, tx=44.0, tn=25.0)
+    insert_quotidienne(dt.date(2019, 7, 25), station_code, tx=42.6, tn=22.0)
 
     ds = TimescaleTemperatureRecordsDataSource()
     req = TemperatureRecordsRequest(
@@ -171,8 +109,8 @@ def test_fetch_records_all_time_returns_entries():
     # Ordre chronologique :
     # 1985-01-16: tx=0.0 → premier record all-time
     # 2019-07-25: tx=42.6 → 42.6 > 0.0 → record progressif
-    insert_quotidienne_full(dt.date(1985, 1, 16), station_code, tx=0.0, tn=-20.5)
-    insert_quotidienne_full(dt.date(2019, 7, 25), station_code, tx=42.6, tn=22.0)
+    insert_quotidienne(dt.date(1985, 1, 16), station_code, tx=0.0, tn=-20.5)
+    insert_quotidienne(dt.date(2019, 7, 25), station_code, tx=42.6, tn=22.0)
 
     ds = TimescaleTemperatureRecordsDataSource()
     req = TemperatureRecordsRequest(period_type="all_time", type_records="hot")
@@ -189,7 +127,7 @@ def test_fetch_records_returns_correct_types():
     station_code = "99005001"
 
     insert_station(station_code, "Station Types Test", department=99)
-    insert_quotidienne_full(dt.date(2019, 7, 25), station_code, tx=42.6, tn=22.0)
+    insert_quotidienne(dt.date(2019, 7, 25), station_code, tx=42.6, tn=22.0)
 
     ds = TimescaleTemperatureRecordsDataSource()
     req = TemperatureRecordsRequest(period_type="all_time", type_records="hot")
